Added train and export scripts
These scripts are provided from TensorFlow
Object Detection API. The train script is used
for training a pretrained model with new data.
So, I grabbed ssd mobilenet and trained this
pretrained model with traffic light images. So,
it already was good at general object detection,
such as identifying those are traffic lights. I
further trained it to identify those are traffic
lights that are red, green, yellow. Then export
inference script, exports the model checkpoint
that was saved from training as a frozen graph,
so we can use it for later. Exporting a frozen graph
is very important for tensorflow version compability.
The ssd mobilenet was trained with tensorflow 1.12,
then I trained it with tensorflow 1.4.0. I used GPU.
The important part, if I want to use it with
tensorflow 1.3.0, then I need to export the trained
model with tensorflow 1.4.0 since they use the same
CUDA and cuDNN versions.