Adjusted updater rate and lookahead wps
I found that the best values for working with the
simulator are lookahead waypoints = 20 and waypoint
updater rate set to 20Hz. I was able to get the car
to drive around in the simulator through most of
the intersections while stopping at red lights.
I did notice sometimes the car would stop at green
lights too, then drive through them when they changed
to yellow, but I figured that was because of latency
that is caused by the camera being turned on in the
simulator.