Brought back camera processing rate
Setting the camera image processing rate
and waypoint threshold actually worked out
better for dealing with the camera latency
coming from the simulator. It didn't solve
the problem, but it helped, so I could create
a video of my traffic light classifier recognizing
the state of traffic lights, stopping at red lights
and driving through green lights when the simulator
did not lag.